ListUsers CLI migration
